// lib/api/generated/schemas/baseSuccessResponseSchema.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Fullstack Auth API
 * Complete authentication and user management API for web and mobile applications. Supports JWT authentication, role-based access control, and user management with caching.
 * OpenAPI spec version: 1.0
 */
import type { BaseSuccessResponseSchemaData } from './baseSuccessResponseSchemaData';

export interface BaseSuccessResponseSchema {
  success: boolean;
  /** Response data */
  data: BaseSuccessResponseSchemaData;
  /** Response timestamp */
  timestamp: string;
}