// lib/api/generated/schemas/validationErrorSchema.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Fullstack Auth API
 * Complete authentication and user management API for web and mobile applications. Supports JWT authentication, role-based access control, and user management with caching.
 * OpenAPI spec version: 1.0
 */
import type { ValidationErrorSchemaError } from './validationErrorSchemaError';

export interface ValidationErrorSchema {
  success: boolean;
  error: ValidationErrorSchemaError;
}